refactor: improve error handling
