+++ src/Eppie.App/Eppie.App.ViewModels/AllMessagesPageViewModel.cs
@@ -10,26 +10,8 @@ namespace Tuvi.App.ViewModels
 {
     public class AllMessagesPageViewModel : MessagesViewModel
     {
-        public class NavigationData
+        public class NavigationData : BaseNavigationData
         {
-            public IErrorHandler ErrorHandler { get; set; }
-        }
-
-        private IErrorHandler PageErrorHandler { get; set; }
-
-        public override void OnNavigatedTo(object data)
-        {
-            if (data is NavigationData navigationData)
-            {
-                PageErrorHandler = navigationData.ErrorHandler;
-            }
-
-            base.OnNavigatedTo(data);
-        }
-
-        public override void OnError(Exception e)
-        {
-            PageErrorHandler?.OnError(e, false);
         }
 
         protected override IEnumerable<MessageInfo> SelectAppropriateMessagesFrom(List<ReceivedMessageInfo> receivedMessages)
